requirements.txt
flask
opencv-python
numpy
google-generativeai
